Huge change: Don't use loaders, but a headless fragment to execute tasks
This helps in reducing code duplication since Loaders were pretty similar.
It also makes it easier to execute tasks (no need to initLoader & restartLoader and forceLoad)
Since the TaskFragment is long-living we can hold an in-memory cache and handle screen rotation gracefully.

Still, not 100% sure if this is a good solution